pyproject.toml
[project]
name = "rovi-ros-ws"
version = "0.1.0"
description = "ROS 2 Jazzy workspace Python dependencies for Room View Bot"
requires-python = ">=3.12"
dependencies = [
    "pyserial>=3.5",
    "psutil>=5.9",
    "grpcio>=1.60",
    "protobuf>=4.25",
    "PyYAML>=6.0",
    "rosmaster-lib",
]

[tool.uv.sources]
rosmaster-lib = { git = "https://github.com/RobLibs/Rosmaster_Lib" }
